Fixes after merge + conversation event helper deleted
